### 2025-06-10 #247 @isaac-aa
MILESpy
Submitting Author: (@isaac-aa)
Package Name: MILESpy
One-Line Description of Package: Python wrapper for the MILES stellar library and Single Stellar Population models
Repository Link:  https://github.com/miles-iac/milespy
Version submitted: 1.0rc3
Version accepted: TBD

---

MILESpy is a python interface to the MILES single stellar population (SSP) models and stellar library. This package aims to provide users an easy interface to access SSP models, navigate the stellar library or synthesize a spectrum given a star formation history (SFH). It automatically downloads all the needed data and includes utilities to post-process the resulting spectra, including computing photometry, rebinning, convolution and velocity shifts.  MILESpy is fully integrated and builds upon previously existing tools, namely astropy and specutils.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

MILESpy has the potential to be used by a wide range of fields in Astronomy and Astrophysics, for example, studies of unresolved stellar populations in galaxies, spectra generation from cosmological simulations, SED fitting. Although this fields have already been using the MILES SSP models and stellar library, MILESpy  aims to reduce development times and increase scientific production by providing a ready-to-use tool for analysis and coupling with third-party codes (e.g., ppxf, PST, FSPS).

  To the best of our knowledge, the [Population Synthesis Toolkit](https://github.com/paranoya/population-synthesis-toolkit) (PST) Python library is the only package that has a similar aim as MILESpy.  For example, it provides an interface to E-MILES SSP models (one of several models in MILESpy), which are also available in MILESpy, but does not provide an interface to any stellar library. Overall, PST focuses on building Composite Stellar Populations (CSP) from different sets of SSP models. Thus, it could potentially even use MILESpy as a back-end to add more SSP models not readily presently in PST.

https://github.com/pyOpenSci/software-submission/issues/237
