### 2025-06-11 fc0eae7
[no test] Update simulation example to use pynbody data
