+++ docs/tutorials/stars.txt
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:f69c380cadefd11c7ef4eabfa6c3f1923ce02d105016d67f3106bbdb56721c90
-size 25255018
