### 2025-07-24 #247 @eliotwrobson
---
---

From the above, my comments are the following:

- You'll need to be in compliance with the README requirements, which you can find [here](https://www.pyopensci.org/python-package-guide/documentation/repository-files/readme-file-best-practices.html) before we can get a full review going.
- It seems like some of the [requirements](https://github.com/miles-iac/milespy/blob/main/pyproject.toml#L34) are pretty out of date? This package doesn't support numpy 2.0. I think this could benefit from an uplift, along with possibly using newer developer tools like uv and ruff (instead of flake8).
- It doesn't look like the existing lint checks run as part of the normal CI? You should add workflows that include this.

Those issues are pretty minor in the scheme of things. Aside from that, I'll check in with others and get the ball rolling 👍🏽
