+++ Makefile
@@ -1,5 +1,5 @@
 .DEFAULT_GOAL := init
-.PHONY: add_major_version add_minor_version add_patch_version add_premajor_version add_preminor_version add_prepatch_version add_prerelease_version prepare-dev install-dev data help lint reference-img tests upload-prod-pypi upload-test-pypi update_req update_req_dev pyclean doc
+.PHONY: install-dev reference-img help lint tests pyclean doc release
 VENV = ".milespy"
 
 define PROJECT_HELP_MSG
@@ -11,36 +11,22 @@ Usage:\n
 	-------------------------------------------------------------------------\n
 	\t\tInstallation\n
 	-------------------------------------------------------------------------\n
-	make\t\t\t\t                Install milespy in the system (root)\n
-	make user\t\t\t 			Install milespy for non-root usage\n
+	make\t\t\t\t                Install milespy in the current environment\n
 	\n
 	-------------------------------------------------------------------------\n
 	\t\tDevelopment\n
 	-------------------------------------------------------------------------\n
-	make prepare-dev\t\t 		Prepare Development environment\n
 	make install-dev\t\t 		Install milespy for development purpose\n
 	make reference-img\t\t     Generate reference images for tests\n
 	make tests\t\t\t             Run units and integration tests\n
 	\n
 	make doc\t\t\t 				Generate the documentation\n
 	\n
-	make release\t\t\t 			Release the package as tar.gz\n
-	\n
-	make update_req\t\t			Update the version of the packages in pyproject.toml\n
+	make release\t\t\t 			Build the distribution files\n
 	\n
 	make pyclean\t\t\t		Clean .pyc files and __pycache__ directories\n
 	\n
-	-------------------------------------------------------------------------\n
-	\t\tVersion\n
-	-------------------------------------------------------------------------\n
-	make version\t\t\t		Display the version\n
-	make add_major_version\t\t	Add a major version\n
-	make add_minor_version\t\t	Add a major version\n
-	make add_patch_version\t\t	Add a major version\n
-	make add_premajor_version\t	Add a pre-major version\n
-	make add_preminor_version\t	Add a pre-minor version\n
-	make add_prepatch_version\t	Add a pre-patch version\n
-	make add_prerelease_version\t	Add a pre-release version\n
+	make envclean\t\t\t		Remove the local development environment\n
 	\n
 	-------------------------------------------------------------------------\n
 	\t\tOthers\n
@@ -56,79 +42,37 @@ export PROJECT_HELP_MSG
 help:
 	echo $$PROJECT_HELP_MSG
 
-
-#
-# Sotware Installation in the system (need root access)
-# -----------------------------------------------------
-#
 init:
-	poetry install --no-dev
-
-#
-# Sotware Installation for user
-# -----------------------------
-# This scheme is designed to be the most convenient solution for users
-# that don’t have write permission to the global site-packages directory or
-# don’t want to install into it.
-#
-user:
-	poetry install --no-dev
-
-prepare-dev:
-	git config --global init.defaultBranch main && git init && echo "python3 -m venv milespy-env && export PYTHONPATH=. && export PATH=`pwd`/milespy-env/bin:${PATH}" > ${VENV} && echo "source \"`pwd`/milespy-env/bin/activate\"" >> ${VENV} && echo "\nnow source this file: \033[31msource ${VENV}\033[0m"
+	python3 -m pip install .
 
 install-dev:
-	poetry install -E docs && poetry run pre-commit install
[…]
