+++ pyproject.toml
@@ -1,64 +1,59 @@
-[tool.poetry]
+[project]
 name = "milespy"
 version = "1.0rc3"
-authors = [
-  "Isaac Alonso Asensio <isaac.alonso.asensio@iac.es>"
-]
-maintainers = [
-  "Isaac Alonso Asensio <isaac.alonso.asensio@iac.es>"
-]
 description = "Wrapper for the MILES spectral library"
+authors = [{ name = "Isaac Alonso Asensio", email = "isaac.alonso.asensio@iac.es" }]
+requires-python = "~=3.8"
 readme = "README.md"
-packages = [{include = "milespy"}]
+maintainers = [{ name = "Isaac Alonso Asensio", email = "isaac.alonso.asensio@iac.es" }]
+dependencies = [
+    "h5py>=3.11.0,<4",
+    "scipy~=1.9 ; python_version < '3.12'",
+    "scipy~=1.10 ; python_version >= '3.12'",
+    "astropy~=5.2 ; python_version < '3.12'",
+    "astropy>=6,<7 ; python_version >= '3.12'",
+    "specutils~=1.15",
+    "spectres>=2.2.2,<3",
+    "setuptools>63 ; python_version >= '3.12'",
+    "matplotlib~=3.7",
+    "numpy~=1.24 ; python_version < '3.12'",
+    "numpy~=1.26 ; python_version >= '3.12'",
+    "requests>=2.31.0,<3",
+    "tqdm>=4.66.4,<5",
+    "rcfile>=0.1.4,<0.2",
+]
+
+[dependency-groups]
+dev = [
+    "black>=24.4.2,<25",
+    "pre-commit~=3.5",
+    "pylint>=3.2.0,<4",
+    "pytest>=8.2.0,<9",
+    "pytest-mpl>=0.17.0,<0.18",
+    "pytest-html>=4.1.1,<5",
+    "coverage>=7.5.4,<8",
+]
+docs = [
+    "Sphinx>=7.1.2,<8",
+    "sphinx-rtd-theme>=2.0.0,<3",
+    "sphinxcontrib-napoleon>=0.7,<0.8",
+    "numpydoc>=1.7.0,<2",
+    "nbsphinx>=0.9.4,<0.10",
+    "sphinx-gallery>=0.17.0,<0.18",
+    "jupyterlab>=4.0.0,<5",
+    "sphinx-autodoc-typehints>=2,<3",
+    "sphinxcontrib-bibtex>=2.6.3,<3",
+]
+
+[tool.hatch.build.targets.sdist]
+include = ["milespy"]
+
+[tool.hatch.build.targets.wheel]
+include = ["milespy"]
 
 [build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
-
-[tool.poetry.dependencies]
-python = "^3.8"
-h5py = "^3.11.0"
-scipy = [
-  {version = "^1.9", python = "<3.12"},
-  {version = "^1.10", python = ">=3.12"}
-]
-astropy = [
-  {version = "^5.2", python = "<3.12"},
-  {version = "^6", python = ">=3.12"}
-]
-specutils = "^1.15"
-spectres = "^2.2.2"
-# setuptools is required by gwcs
-setuptools = {version = ">63", python = ">=3.12"}
-matplotlib = "^3.7"
-numpy = [
-  {version = "^1.24", python = "<3.12"},
-  {version = "^1.26", python = ">=3.12"}
-]
-requests = "^2.31.0"
-tqdm = "^4.66.4"
-rcfile = "^0.1.4"
-Sphinx = { version = "^7.1.2", optional = true }
-sphinx-rtd-theme = { version = "^2.0.0", optional = true }
[…]
