+++ .pre-commit-config.yaml
@@ -22,12 +22,9 @@ repos:
     rev: v3.12.0
     hooks:
       - id: reorder-python-imports
-  - repo: https://github.com/psf/black.git
-    rev: 23.10.1
+  - repo: https://github.com/astral-sh/ruff-pre-commit
+    rev: "v0.12.5"
     hooks:
-      - id: black
-        language_version: python3
-  - repo: https://github.com/pycqa/flake8
-    rev: 6.1.0
-    hooks:
-      - id: flake8
+      - id: ruff
+        args: ["--fix", --max-line-lenght=88]
+      - id: ruff-format
