### 2025-07-28 #247 @isaac-aa
Hello, thanks for the comments. I have updated the README to add links to the documentation and I think that now it adheres to the requirements that you linked.

Regarding to the use of other developers tools (uv and ruff), I have started the migration, as I'm already using them in other projects. The linting will also be added to the CI once this is done.

The requirement for numpy < 2.0 has been decided to allow for older versions of python to run milespy, as some packages for python 3.8 do not support numpy >= 2.0. In any case, I could check if this is still the case and update the requirements if I can.
