### 2025-07-29 19349bb
Small linting fix after ruff transition
