### 2025-07-29 a42e61a
Add ruff in the github workflow
