+++ pyproject.toml
@@ -32,6 +32,7 @@ dev = [
     "pytest-mpl>=0.17.0,<0.18",
     "pytest-html>=4.1.1,<5",
     "coverage>=7.5.4,<8",
+    "ruff>=0.12.5",
 ]
 docs = [
     "Sphinx>=7.1.2,<8",
