### 2025-07-29 c09ccae
Add linting/formatting step in CI
