+++ .github/workflows/test.yml
@@ -7,6 +7,15 @@ concurrency:
   cancel-in-progress: true
 
 jobs:
+  lint-format:
+    runs-on: ubuntu-22.04
+    steps:
+      - run: sudo apt -y update
+      - uses: actions/checkout@v4
+      - name: Setup the environment
+        uses: ./.github/actions/setup
+      - run: ruff check
+      - run: ruff format
   test:
     permissions:
       contents: write
