### 2025-07-29 27cc9fb
Add Makefile changes to use ruff
