+++ .github/actions/setup/action.yaml
@@ -10,10 +10,7 @@ runs:
     - uses: actions/setup-python@v5
       with:
         python-version: ${{ inputs.python-version }}
-    - name: Install ruff
-      uses: astral-sh/ruff-action@v3
-      with:
-        version-file: "pyproject.toml"
-        args: "--version"
+    - name: Install uv
+      uses: astral-sh/setup-uv@v6
     - run: make install-dev
       shell: bash
