### 2025-07-29 175cdd4
Update dev docs to use uv
