+++ docs/development.rst
@@ -24,17 +24,16 @@ But if you want to develop and contribute changes to MILESpy, the
 best way is to use the provided
 `Makefile <https://github.com/miles-iac/milespy/blob/main/Makefile>`_
 to setup the environment.
-Under the hood, it uses `poetry <https://python-poetry.org/>`_
+Under the hood, it uses `uv <https://docs.astral.sh/uv/>`_
 for environment management and dependency solving, so you need to
-`install it <https://python-poetry.org/docs/#installation>`_.
+`install it <https://docs.astral.sh/uv/getting-started/installation/>`_.
 
 Then, you can install MILESpy and all the required development dependencies:
 
 .. code-block:: bash
 
-  make prepare-dev
-  . .milespy
   make install-dev
+  . .venv/bin/activate
 
 After you finish adding new functionalities, do not forget to run the tests
 to check that your contribution does not break working bits of MILESpy. Also,
