### 2025-07-29 7515f4b
Updated requirements and drop support for python < 3.10
This allows for using specutils>=2, numpy>2 and astropy>7
