+++ .github/workflows/test.yml
@@ -24,11 +24,10 @@ jobs:
     strategy:
       matrix:
         python-version:
-        - '3.8'
-        - '3.9'
         - '3.10'
         - '3.11'
         - '3.12'
+        - '3.13'
     steps:
     - run: sudo apt -y update
     - uses: actions/checkout@v4
