### 2025-07-29 #247 @isaac-aa
We have updated the requirements (and dropped support for python<3.10) so that we are compatible with the latest versions of our most critical dependencies, namely `numpy`, `astropy` and `specutils`.
