### 2025-11-03 9f9b177
Add sMILES SSP
