+++ milespy/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-__version__ = "1.0rc3"
+__version__ = "1.0rc5"
 
 
 from .ssp_library import SSPLibrary
