### 2026-02-01 #247 @crhea93
Hi @isaac-aa I'll be stepping in as the editor in chief for this submission. I'll continuing @eliotwrobson's excellent work in finding an editor. Please feel free to reach out if anything comes up and thank you for your patience!
