### 2026-02-11 #247 @hamogu
Hi @isaac-aa ! I'll be acting as editor for this submission and I'm looking for reviewers right now.
I'm sorry it took so long - we have more submissions at PyOpenSci than volunteer editors and reviewers. You might think that's a good problem to have, but unfortunately that means that we're far slower with the reviews than we want to be. Sorry!

Anyway, I hope I'll find reviewers soon and then we're on our way.
