### 2026-03-10 #247 @rosteen
- [x] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
- [x] **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
- [x] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

The package structure should follow the general community best practices. In general, please consider whether:

---

I've started this review and will be checking things off as I work through it. A couple of initial comments:

1. In the `Configuration` section of the docs, it would be nice to provide an example of the expected format of the configuration file (JSON? YAML?) in addition to specifying where it can be stored.
2. I'm running the first tutorial's code now - when I reached the prompt `Do you want to download the MILES_SSP_v9.1 repository? [y/n]:` I accidentally hit enter without entering anything, which led to a traceback (ending in `AssertionError: Repository file is unreadable`). I think it would be better to re-prompt the user if something other than `y` or `n` is input rather than throwing an error.
3. Running `spec.mass_to_light` in the tutorial code leads to an error traceback, the root issue being that it uses `np.trapz` which no longer exists in numpy 2.x:
  ```
    File "<python-input-25>", line 3, in <module>
      outmls = spec.mass_to_light(filters=filts, mass_in="star+remn")
    File "/Users/rosteen/projects/milespy/milespy/spectra.py", line 501, in mass_to_light
      mags = self.magnitudes(filters=filters, zeropoint=zeropoint)
    File "/Users/rosteen/projects/milespy/milespy/spectra.py", line 384, in magnitudes
      outmags = compute_mags(self.spectral_axis, self.flux, filters, zeropoint)
    File "/Users/rosteen/projects/milespy/milespy/magnitudes.py", line 140, in compute_mags
      vega_f = np.trapz(vega * response, x=tmp_wave)
               ^^^^^^^^
    File "/Users/rosteen/opt/anaconda3/envs/milespy_testing/lib/python3.13/site-packages/numpy/__init__.py", line 792, in __getattr__
      raise AttributeError(f"module {__name__!r} has no attribute {attr!r}")
  AttributeError: module 'numpy' has no attribute 'trapz'. Did you mean: 'trace'?
  ```

#### Additional Review Comments 2026-03-11

- I didn't see any citation information, I would recommend a CITATION.cff file (see astropy's [here](https://github.com/astropy/astropy/blob/main/CITATION.cff)) unless I missed this information somewhere.
- It might be nice to have a list_all (or whatever name makes sense) method for filters and indices. You can do this via search() but it's a little awkward having to provide an empty string as input (calling with no argument errors).
- Related, maybe have a docs section dedicated to filters and ls_indices, rather than having that functionality interspersed with the other tutorial sections? Not something I'd demand be addressed, just a thought I had as I was going through the code.
- Should ls_indices be in the API docs under "other tools"?
- Have you considered using the specutils [function for converting between air and vacuum wavelengths](https://github.com/astropy/specutils/blob/main/specutils/utils/wcs_utils.py), which gives the option to choose between multiple methods? (Instead of magnitudes.vacuum2air)
