### 2026-03-11 #247 @rosteen
@isaac-aa FYI I edited the main review comment above to leave some additional comments. Overall this seems like a very useful package and I didn't see any big issues with the code while I was reading through it! I still need to do some more testing locally just to make sure everything is working. I did see you pushed a commit fixing the numpy issue already, thanks.
