### 2026-03-11 #247 @A-Derks
- [x] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
- [x] **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
	**_Needs more documentation and examples_**
- [ ] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.
	**_Please add project urls- repository url, documentation url, possibly a homepage url which would link to MILES homepage_**

	**_Please add python versions supported badge_**

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

	**_Are there other packages to compare MILESPy to?_**
	**_Please add citation information_**

The package structure should follow the general community best practices. In general, please consider whether:

	**_Please make tests automated. A user should be able to run the entire test suite from the terminal. I prefer pytest, so all test files should start with ‘test_’ in order for pytest to recognize and run the tests.
	There are options other than pytest though! Along the same lines, all of the .png files should not be included in this repo. Testing should test the functionality and reproducibility of the package._**
	**_Please bump up from 3.10_**
	**_CI tests are passing but are raising some warnings- please check these out_**

**_I don’t think you are submitting to JOSS so I will not review this section_**

5 h

Package currently supports Python 3.10+. I would suggest bumping this to 3.11 or 3.12 as support for 3.10 and 3.11 are reaching end of life.
I would suggest including a changelog in the repo and also in the documentation.
It would be helpful to add a simple, one line comment at the top of each module to give a quick overview of what the module does.
filter.py:
	It is a bit confusing that there are two plot functions in one module. Filter.plot() and plot(). I would suggest changing the names of these functions to be a bit more descriptive. I.e. Filter.plot_filter_transmissivity() and plot_filter().
	I also suggest doing the same for search() and get(). I.e. filter_search() and get_filter() or something similar.
Ls_indices.py:
	Instead of using ‘band_blue1’, ‘band_blue2’, etc. it would be more readable if these were ‘band_blue_left_edge’, ’band_blue_right_edge’, etc.
	Consider using Pydantic to validate the attributes in here.
	Please add documentation to the LineStrengthIndex __init__
	I would suggest making the function names in here more descriptive of what they do. I.e. get() —> get_index_from_database(), search() —> search_for_index_in_database(). That way, a user does not have to reference the function documentation to see what it does, it is just readable!
	lsindex() —> line_strength_index()
	Address TODO in lsindex() function!
	Remove commented out code at the end of lsindex() function or uncomment it and make it work!
Magnitudes.py:
	Use Pydantic to verify that zeropoint attr only ever gets ‘AB’ or ‘VEGA’
	Why are there question marks in sun attr documentation? Are you sure about False
	Address TODO in here
	line 101 typo in ‘Angstrom’
	vacuum2tair documentation attr doesn’t match actual attr
	compute_mags() function is a really long and hard to read function. It should have some helper functions. I.e. everywhere you have a comment and a block of code, that can be its own function.
	Ex:
		# Computing the magnitude for each filter
			for filt in filters:
    		# Finding the wavelength limits of the filters
    			good = filt.wave > 0.0
    			wlow = np.amin(filt.wave[good])
    			whi = np.amax(filt.wave[good])
	can be all put into a single helper function called find_wavelength_limits()
Misc.py
	This module needs to be renamed. It is not helpful.
Repository.py
	Can you add documentation in here? I am not sure what this module is or does.
SFH.py
	This module and its class needs to be renamed. SFH is jargon. Please name to StarFormationHistories
	Same with other attributes in this class. SFR, met, imf. Please rename. They are all jargon.
spectra.py
	The filterwarnings is a red flag to me. Why are you ignoring those?
	Please add documentation to all properties
	Remove commented out code in this module
	Please rename all jargon acronyms in here too. I.e. lsf
	This module seems like it is still under construction….
ssp_library.py
	This module and its class needs to be renamed. SSP is jargon. Please name to SingleStellarPopulation
	Again, why do you have a filterwarnings in here?
	Use Pydantic to only accept the correct values of isochrone
	__Init__ could be split into helper functions to make the code more readable
	super().__init__ should be closer to top of __init__
	Please fix jargon variables in here
	Most of the functions in here are super long and need to be split into multiple other functions
