### 2026-03-17 #247 @isaac-aa
Hello,
First, thanks both @A-Derks and @rosteen for the detailed and constructive review of MILESpy!
I have updated the `main` branch to incorporate the changes due to this review process.

I will address the main comments here for each reviewer:
# @rosteen
## **Review comments**:
1. I have added the format in the documentation and a small example.
2. That behaviour is intended, if the user does not respond with `y`, the repository
    files will not be downloaded and thus will not be available to MILESpy, causing
    the `AssertionError: Repository file is unreadable`. To make this more clear
    to the user, I have added a warning in the case that the response is not `y`.
3. I have updated the use of `np.trapz` to `np.trapezoid`. This also bumped the
    required numpy version to be `>=2`.
## **Additional comments**
1. We have not included a `CITATION.cff` because we still not have a publish paper
    about MILESpy. The paper will soon be submitted, and then added as you suggested.
    In any case, we provide in the documentation the references of the libraries
    that are used in the package for their correct citation if needed.
2. I have added `list_all` options for both filters and line strength indices.
3. I have added the line strength indices documentation directly in "other tools".
4. I was not aware of the `vac_to_air` function of `specutils`. We now use that.

# @A-Derks
1. We want to keep providing support for python 3.10. The reason being that in
   our institution (where most of the current users of MILESpy are), the default
   OS version is Ubuntu 22.04, which still ships python 3.10 by default. The EOL
   cycle of that Ubuntu version is on 2027. Until then, we plan to maintain
   support to python 3.10 if there is no technical reason to drop it.
2. I have included a changelog both in the repository and in the documentation
3. I have added a one-liner describing the modules content
## **filter.py**
1. I have clarified the plot function names. Now there is a general `plot_filters`
   that given a list of filter plots their transmissivity using the `plot_transmissivity`
   method inside the `Filter` class.
2. I have **not** modified the names `get()`->`get_filter()` as I believe that it
   is too verbose. The same applies for the line strength indices. As show in the
   examples/tutorials, to use that functions the recommended way is:
   ```python
   import milespy.filters as flib
   flib.get(...)
   ```
   I think that `flib.get_filters()` does not add any useful information to the users.
   The same applies to the other functions in the module.
## **ls_indices.py**
1. I have improved the naming of the `band_blue1` and friends.
2. We now use pydantic to validate the inputs
3. As with `filter.py`, we have decided to keep the short functions names
4. Overall we have cleaned, commented and improve this module. In addition, we
   have renamed it to `line_strength_indices.py` to avoid jargon.
## **magnitudes.py**
1. We now use pydantic to validate the inputs
2. The TODO and `sun` attrs have been fixed/clarified
3. Fixed typo in Angstrom
4. The `vacuum2air` function has been eliminated and now we use `specutils`'s `vac_to_air`.
5. We have refactored `compute_mags()` to reduce its complexity
## **misc.py**
1. The module has been renamed to `interpolation_utils.py` to clarify its scope.
## **repository.py**
1. Clarifying documentation has been added both at the module and class level.
## **SFH.py**
1. We have renamed the module to `star_formation_history.py` and the main class
   to `StarFormationHistory`. However, due to its shortness and its ubiquitous
   use in the community, we have left `SFH` as an alias.
2. We have refrain to change significantly the API of this and other portions
   of the package as we believe that a user of MILESpy will have basic knowledge
   of astrophysics, which involves the common use of acronyms for star formation
   histories, single stellar populations, and so on. In some cases, we have defined
   long-names without jargon and short acronyms as aliases for cleaner user code.
## **spectra.py**
1. We have removed the `filterwarnings`
2. The have cleaned and commented the properties and methods of the main class.
## **ssp_library.py**
1. See point 1. of **SFH.py**
2. Added pydantic and removed filterwarnings
3. We have reduced the complexity of the `init` function and the rest.
   However, the place of `super().init` can not be changed due to technical constraints.
