### 2026-04-02 #247 @rosteen
For what it's worth, you don't need to have submitted a paper to have a `CITATION.cff` file, you can always create a Zenodo record and associated DOI for the package. See for example https://github.com/spacetelescope/jdaviz/blob/main/CITATION.cff and  https://doi.org/10.5281/zenodo.5513927.
