### 2026-04-03 #247 @A-Derks
This is looking REALLY good!!! Thanks for all of your hard work and changes @isaac-aa!

- [x] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
- [x] **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
- [x] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.

	**_Please add python versions supported badge_**

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

	**_Are there other packages to compare MILESPy to?_**

The package structure should follow the general community best practices. In general, please consider whether:

**_I don’t think you are submitting to JOSS so I will not review this section_**

6h

#### Review Comments (just a few minor things!)
1. Please add python versions supported badge to README
2. I now understand why you have all of the .png files in the test file. However, I am wondering if those should not be included in the repo, and a user should run `pytest --mpl-generate-path=baseline` to make the baseline images before running the tests. I am really not sure. Is it typical for the developer to include the baseline images?
3. I am getting some failing tests in test/ssp_test.py and test/stellar_library_test.py because I do not have all of the data downloaded. You could host your data on Zenodo, and then use the pooch library in your code and tests to go fetch the data so that a user does not encounter this road block.
