### 2026-04-03 #247 @rosteen
I noticed a couple small problems in the docs while re-reviewing:

1. In `tutorials/index.rst`, "manipulation" is misspelled as "manitulation"
2. In `tutorials/example_ssp.ipynb`, the example still tries to import `ls_indices` rather than the updated module name.
