### 2026-04-22 #247 @isaac-aa
Hello,

I was waiting from some other collaborators the "ok" for putting the file on Zenodo, in the following days I will do that and fix the other small issues.
